refactor(login): extract write-target decision and gitignore helpers
Pulls promptLocation, the --local/--global decision tree, and the
.gitignore prompt out of login.ts so workspace set can reuse them. No
behavior change.
